+++ src/components/Navbar.js
@@ -1,7 +1,7 @@
 // Navbar.js
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './Navbar.css'; // Assurez-vous que ce fichier est correctement importé
+import '../styles/Navbar.css'; 
 
 function Navbar() {
   return (
