+++ src/pages/Home.js
@@ -1,0 +1,15 @@
+// Dino.js
+import logo from '../styles/logo.svg';
+
+
+const Dino = () => {
+
+  return (
+    <div>
+      <img src={logo} className="App-logo" alt="logo" />
+      <h2>Page Home</h2>
+    </div>
+  );
+};
+
+export default Dino;
